# perf(web): use DOMString for BlobParts