70% faster than Node (24ms)
50% faster than Bun (14ms)

(above numbers taken from https://twitter.com/jarredsumner/status/1503888103227867136)

Before:
```
blob_text_large:     	n = 3, dt = 2.123s, r = 1/s, t = 707666666ns/op
```

After:
```
blob_text_large:     	n = 3, dt = 0.021s, r = 143/s, t = 7000000ns/op
```

Closes #13973